+++ FA2sp/Ext/CMapValidator/Body.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include <CMapValidator.h>
+
+class CMapValidatorExt : public CMapValidator
+{
+public:
+	ppmfc::CString FetchLanguageString(const char* Key, const char* def);
+	void InsertString(const char* String, bool IsWarning);
+};