+++ FA2sp/FA2sp.Constants.h
@@ -4,8 +4,8 @@
 #define __str_(x) #x
 
 #define PRODUCT_MAJOR 1
-#define PRODUCT_MINOR 5
-#define PRODUCT_REVISION 2
+#define PRODUCT_MINOR 6
+#define PRODUCT_REVISION 0
 
 #ifdef NDEBUG
 #define PRODUCT_STR __str(PRODUCT_MAJOR) "." __str(PRODUCT_MINOR) "." __str(PRODUCT_REVISION)